Customers can no longer be deleted if they have previously placed orders

+++ Krupa's Foods/Primary Views/CustomersView.swift
@@ -12,6 +12,8 @@ struct CustomersView: View {
     @Query var customers: [Customer]
     @Environment(\.modelContext) var modelContext
     
+    @State private var presentCannotDeleteAlert = false
+    
     var body: some View {
         NavigationStack {
             List {
@@ -26,13 +28,25 @@ struct CustomersView: View {
             .padding(.top, 65)
             #endif
             .navigationTitle("Customers")
+            .alert("This customer cannot be deleted.", isPresented: $presentCannotDeleteAlert) {
+                Button("OK", role: .cancel) {
+                    presentCannotDeleteAlert = false
+                }
+            } message: {
+                Text("This customer has previously placed orders. Delete associated orders to delete this customer.")
+            }
+
         }
     }
     
     private func deleteCustomer(at offsets: IndexSet) {
         for index in offsets {
             let customer = customers[index]
-            modelContext.delete(customer)
+            if customer.wrappedOrderHistory.isEmpty {
+                modelContext.delete(customer)
+            } else {
+                presentCannotDeleteAlert = true
+            }
         }
         
         try? modelContext.save()
